[Project][Tasks] STATUSES and board removed from store
