+++ urbanvitaliz/frontend/src/js/config/statuses.js
@@ -1,0 +1,10 @@
+export const STATUSES = {
+    PROPOSED: 0,
+    INPROGRESS: 1,
+    BLOCKED: 2,
+    DONE: 3,
+    NOT_INTERESTED: 4,
+    ALREADY_DONE: 5,
+}
+
+export default STATUSES
